# asm: optimize node allocation