--- a/internal/asm/arm64/impl.go
+++ b/internal/asm/arm64/impl.go
@@ -201,7 +201,7 @@ const (
 
 // AssemblerImpl implements Assembler.
 type AssemblerImpl struct {
-	nodePool *nodePool
+	nodePool nodePool
 	asm.BaseAssemblerImpl
 	root, current     *nodeImpl
 	buf               *bytes.Buffer
@@ -214,37 +214,51 @@ type AssemblerImpl struct {
 	relativeJumpNodes, adrInstructionNodes []*nodeImpl
 }
 
-const nodePoolPageSize = 1000
+const nodePageSize = 128
+
+type nodePage = [nodePageSize]nodeImpl
 
 // nodePool is the central allocation pool for nodeImpl used by a single AssemblerImpl.
 // This reduces the allocations over compilation by reusing AssemblerImpl.
 type nodePool struct {
-	pages [][nodePoolPageSize]nodeImpl
-	// page is the index on pages to allocate node on.
-	page,
-	// pos is the index on pages[.page] where the next allocation target exists.
-	pos int
+	pages []*nodePage
+	index int
 }
 
 // allocNode allocates a new nodeImpl for use from the pool.
 // This expands the pool if there is no space left for it.
-func (n *nodePool) allocNode() (ret *nodeImpl) {
-	if n.pos == nodePoolPageSize {
-		if len(n.pages)-1 == n.page {
-			n.pages = append(n.pages, [nodePoolPageSize]nodeImpl{})
+func (n *nodePool) allocNode() *nodeImpl {
+	if n.index == nodePageSize {
+		if len(n.pages) == cap(n.pages) {
+			n.pages = append(n.pages, new(nodePage))
+		} else {
+			i := len(n.pages)
+			n.pages = n.pages[:i+1]
+			if n.pages[i] == nil {
+				n.pages[i] = new(nodePage)
+			}
 		}
-		n.page++
-		n.pos = 0
+		n.index = 0
 	}
-	ret = &n.pages[n.page][n.pos]
-	*ret = nodeImpl{}
-	n.pos++
-	return
+	ret := &n.pages[len(n.pages)-1][n.index]
+	n.index++
+	return ret
+}
+
+func (n *nodePool) reset() {
+	for _, ns := range n.pages {
+		pages := ns[:]
+		for i := range pages {
+			pages[i] = nodeImpl{}
+		}
+	}
+	n.pages = n.pages[:0]
+	n.index = nodePageSize
 }
 
 func NewAssembler(temporaryRegister asm.Register) *AssemblerImpl {
 	return &AssemblerImpl{
-		nodePool:                       &nodePool{pages: [][nodePoolPageSize]nodeImpl{{}}},
+		nodePool:                       nodePool{index: nodePageSize},
 		buf:                            bytes.NewBuffer(nil),
 		temporaryRegister:              temporaryRegister,
 		pool:                           asm.NewStaticConstPool(),
@@ -267,12 +281,13 @@ func (a *AssemblerImpl) Add(n asm.Node) {
 
 // Reset implements asm.AssemblerBase.
 func (a *AssemblerImpl) Reset() {
-	buf, np, tmp := a.buf, a.nodePool, a.temporaryRegister
 	pool := a.pool
 	pool.Reset()
 	*a = AssemblerImpl{
-		buf: buf, nodePool: np, pool: pool,
-		temporaryRegister:   tmp,
+		buf:                 a.buf,
+		nodePool:            a.nodePool,
+		pool:                pool,
+		temporaryRegister:   a.temporaryRegister,
 		adrInstructionNodes: a.adrInstructionNodes[:0],
 		relativeJumpNodes:   a.relativeJumpNodes[:0],
 		BaseAssemblerImpl: asm.BaseAssemblerImpl{
@@ -280,8 +295,8 @@ func (a *AssemblerImpl) Reset() {
 			JumpTableEntries:           a.JumpTableEntries[:0],
 		},
 	}
-	a.nodePool.pos, a.nodePool.page = 0, 0
 	a.buf.Reset()
+	a.nodePool.reset()
 }
 
 // newNode creates a new Node and appends it into the linked list.